+++ tests/e2e/cases/symbol_bigint_object_methods/in.ts
@@ -14,3 +14,4 @@ console.log("big-locale:", big.toLocaleString());
 console.log("big-value:", big.valueOf() + 1n);
 console.log("own:", sym.hasOwnProperty("description"), big.propertyIsEnumerable("toString"));
 console.log("ignored big:", BigInt("255", mark("b")).toString(16, mark("s")), big.toLocaleString(mark("l")), big.valueOf(mark("v")) === big, seen);
+console.log("ignored own:", sym.hasOwnProperty("description", mark("h")), big.propertyIsEnumerable("toString", mark("e")), seen);